# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Furniture store API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Product operations
  - name: categories
    description: Category operations
  - name: orders
    description: Order operations
  - name: offers
    description: Offer operations
  - name: portfolio
    description: Portfolio operations
  - name: testimonials
    description: Testimonial operations
  - name: contact
    description: Contact operations
  - name: stats
    description: Stats and summaries
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: sort
          in: query
          schema:
            type: string
            enum: [newest, price_asc, price_desc]
        - name: featured
          in: query
          schema:
            type: boolean
        - name: offer
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create product (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

  /products/featured:
    get:
      operationId: getFeaturedProducts
      tags: [products]
      summary: Get featured products for homepage
      responses:
        "200":
          description: Featured products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List categories
      responses:
        "200":
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create category (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"

  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders (admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, delivered, cancelled]
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /offers:
    get:
      operationId: listOffers
      tags: [offers]
      summary: List active offers
      responses:
        "200":
          description: Offers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Offer"
    post:
      operationId: createOffer
      tags: [offers]
      summary: Create offer (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OfferInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Offer"

  /offers/{id}:
    patch:
      operationId: updateOffer
      tags: [offers]
      summary: Update offer (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OfferUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Offer"
    delete:
      operationId: deleteOffer
      tags: [offers]
      summary: Delete offer (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /portfolio:
    get:
      operationId: listPortfolio
      tags: [portfolio]
      summary: List portfolio items
      responses:
        "200":
          description: Portfolio items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PortfolioItem"
    post:
      operationId: createPortfolioItem
      tags: [portfolio]
      summary: Create portfolio item (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PortfolioInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortfolioItem"

  /testimonials:
    get:
      operationId: listTestimonials
      tags: [testimonials]
      summary: List testimonials
      responses:
        "200":
          description: Testimonials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Testimonial"
    post:
      operationId: createTestimonial
      tags: [testimonials]
      summary: Create testimonial (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestimonialInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Testimonial"

  /contact:
    post:
      operationId: submitContact
      tags: [contact]
      summary: Submit contact form
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Contact submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactMessage"

  /stats/summary:
    get:
      operationId: getStatsSummary
      tags: [stats]
      summary: Get admin stats summary
      responses:
        "200":
          description: Stats summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatsSummary"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    Product:
      type: object
      required: [id, name, price, categoryId, images, isFeatured, isOffer, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
        categoryId:
          type: integer
        categoryName:
          type: ["string", "null"]
        images:
          type: array
          items:
            type: string
        sizes:
          type: ["string", "null"]
        materials:
          type: ["string", "null"]
        colors:
          type: ["string", "null"]
        executionDays:
          type: ["integer", "null"]
        isFeatured:
          type: boolean
        isOffer:
          type: boolean
        discountPercent:
          type: ["integer", "null"]
        createdAt:
          type: string

    ProductInput:
      type: object
      required: [name, price, categoryId]
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        originalPrice:
          type: number
        categoryId:
          type: integer
        images:
          type: array
          items:
            type: string
        sizes:
          type: string
        materials:
          type: string
        colors:
          type: string
        executionDays:
          type: integer
        isFeatured:
          type: boolean
        isOffer:
          type: boolean
        discountPercent:
          type: integer

    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        originalPrice:
          type: number
        categoryId:
          type: integer
        images:
          type: array
          items:
            type: string
        sizes:
          type: string
        materials:
          type: string
        colors:
          type: string
        executionDays:
          type: integer
        isFeatured:
          type: boolean
        isOffer:
          type: boolean
        discountPercent:
          type: integer

    Category:
      type: object
      required: [id, name, slug, productCount]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        icon:
          type: ["string", "null"]
        image:
          type: ["string", "null"]
        productCount:
          type: integer

    CategoryInput:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
        slug:
          type: string
        icon:
          type: string
        image:
          type: string

    Order:
      type: object
      required: [id, customerName, phone, governorate, address, items, total, status, createdAt]
      properties:
        id:
          type: integer
        customerName:
          type: string
        phone:
          type: string
        governorate:
          type: string
        address:
          type: string
        notes:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        total:
          type: number
        status:
          type: string
          enum: [pending, confirmed, delivered, cancelled]
        createdAt:
          type: string

    OrderItem:
      type: object
      required: [productId, productName, quantity, price]
      properties:
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        price:
          type: number
        image:
          type: ["string", "null"]

    OrderInput:
      type: object
      required: [customerName, phone, governorate, address, items]
      properties:
        customerName:
          type: string
        phone:
          type: string
        governorate:
          type: string
        address:
          type: string
        notes:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemInput"

    OrderItemInput:
      type: object
      required: [productId, quantity, price]
      properties:
        productId:
          type: integer
        quantity:
          type: integer
        price:
          type: number

    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, confirmed, delivered, cancelled]

    Offer:
      type: object
      required: [id, title, description, discountPercent, isActive, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        discountPercent:
          type: integer
        image:
          type: ["string", "null"]
        isActive:
          type: boolean
        validUntil:
          type: ["string", "null"]
        createdAt:
          type: string

    OfferInput:
      type: object
      required: [title, description, discountPercent]
      properties:
        title:
          type: string
        description:
          type: string
        discountPercent:
          type: integer
        image:
          type: string
        isActive:
          type: boolean
        validUntil:
          type: string

    OfferUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        discountPercent:
          type: integer
        image:
          type: string
        isActive:
          type: boolean
        validUntil:
          type: string

    PortfolioItem:
      type: object
      required: [id, title, images, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        images:
          type: array
          items:
            type: string
        beforeImages:
          type: array
          items:
            type: string
        videoUrl:
          type: ["string", "null"]
        customerName:
          type: ["string", "null"]
        createdAt:
          type: string

    PortfolioInput:
      type: object
      required: [title, images]
      properties:
        title:
          type: string
        description:
          type: string
        images:
          type: array
          items:
            type: string
        beforeImages:
          type: array
          items:
            type: string
        videoUrl:
          type: string
        customerName:
          type: string

    Testimonial:
      type: object
      required: [id, customerName, rating, comment, createdAt]
      properties:
        id:
          type: integer
        customerName:
          type: string
        rating:
          type: integer
        comment:
          type: string
        image:
          type: ["string", "null"]
        createdAt:
          type: string

    TestimonialInput:
      type: object
      required: [customerName, rating, comment]
      properties:
        customerName:
          type: string
        rating:
          type: integer
        comment:
          type: string
        image:
          type: string

    ContactInput:
      type: object
      required: [name, phone, message]
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        message:
          type: string

    ContactMessage:
      type: object
      required: [id, name, phone, message, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        message:
          type: string
        createdAt:
          type: string

    StatsSummary:
      type: object
      required: [totalProducts, totalOrders, pendingOrders, totalRevenue, totalCategories]
      properties:
        totalProducts:
          type: integer
        totalOrders:
          type: integer
        pendingOrders:
          type: integer
        totalRevenue:
          type: number
        totalCategories:
          type: integer