# lib/api-zod/src/generated/types/contactInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Furniture store API
 * OpenAPI spec version: 0.1.0
 */

export interface ContactInput {
  name: string;
  phone: string;
  email?: string;
  message: string;
}