# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Furniture store API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  customerName: string;
  phone: string;
  governorate: string;
  address: string;
  /** @nullable */
  notes?: string | null;
  items: OrderItem[];
  total: number;
  status: OrderStatus;
  createdAt: string;
}