# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Furniture store API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  categoryId: number;
  /** @nullable */
  categoryName?: string | null;
  images: string[];
  /** @nullable */
  sizes?: string | null;
  /** @nullable */
  materials?: string | null;
  /** @nullable */
  colors?: string | null;
  /** @nullable */
  executionDays?: number | null;
  isFeatured: boolean;
  isOffer: boolean;
  /** @nullable */
  discountPercent?: number | null;
  createdAt: string;
}