# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Furniture store API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  name: string;
  description?: string;
  price: number;
  originalPrice?: number;
  categoryId: number;
  images?: string[];
  sizes?: string;
  materials?: string;
  colors?: string;
  executionDays?: number;
  isFeatured?: boolean;
  isOffer?: boolean;
  discountPercent?: number;
}