# lib/api-zod/src/generated/types/testimonial.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Furniture store API
 * OpenAPI spec version: 0.1.0
 */

export interface Testimonial {
  id: number;
  customerName: string;
  rating: number;
  comment: string;
  /** @nullable */
  image?: string | null;
  createdAt: string;
}